Migrate reach-related functions to new BlockRayTrace util class
